requirements.txt
# Core AI and ML frameworks
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langchain-tavily>=0.1.0
langchain-chroma>=0.1.0
langgraph>=0.1.0
langgraph-checkpoint-sqlite>=1.0.0

# Web framework
streamlit>=1.28.0

# Data processing
pandas>=2.0.0

# Configuration and validation
pydantic>=2.5.0
python-dotenv>=1.0.0

# Graph visualization
graphviz>=0.20.0
pygraphviz>=1.10.0

# Development tools
ipython>=8.15.0

# Optional: For enhanced performance
# watchdog>=3.0.0  # Streamlit file watching